refactor(metrics): extract computeAggregateMetrics to metrics-aggregation.ts
Extract the pure-function aggregation logic (163 lines) into
metrics-aggregation.ts, reducing metrics.ts from 636 → 475 lines.

The class MetricsCollector (378 lines) is now under the 500-line
quality gate. Re-exported from metrics.ts for backward compatibility
— all consumers import from metrics.js continue to work unchanged.

All 42 metrics/aggregation tests pass.

Architecture Fix Plan: metrics.ts now under 500-line gate ✅
